os-creative-api/requirements.txt
flask 
pillow 
requests 
rembg 
google-cloud-storage
onnxruntime
google-cloud-vision
together
webcolors
replicate